+++ index.js
@@ -1,10 +1,6 @@
 // Components
 export { default as ActionMenu } from './lib/ActionMenu';
 export { default as AlternativeNamesFieldArray } from './lib/AlternativeNamesFieldArray';
-export { default as Card } from './lib/Card';
-export { default as CustomPropertiesConfigListFieldArray } from './lib/CustomPropertiesConfigListFieldArray';
-export { default as CustomPropertiesList } from './lib/CustomPropertiesList';
-export { default as CustomPropertyFilters } from './lib/CustomPropertyFilters';
 export { default as DocumentCard } from './lib/DocumentCard';
 export { default as DocumentsFieldArray } from './lib/DocumentsFieldArray';
 export { default as DuplicateModal } from './lib/DuplicateModal';
@@ -14,7 +10,6 @@ export { default as EResourceType } from './lib/EResourceType';
 export { default as FileUploader } from './lib/FileUploader';
 export { default as FileUploaderField } from './lib/FileUploaderField';
 export { default as FormattedDateTime } from './lib/FormattedDateTime';
-export { default as FormCustomProperties } from './lib/FormCustomProperties';
 export { default as InternalContactCard } from './lib/InternalContactCard';
 export { default as InternalContactSelection } from './lib/InternalContactSelection';
 export { default as InternalContactsFieldArray } from './lib/InternalContactsFieldArray';
@@ -37,7 +32,6 @@ export { default as withAsyncValidation } from './lib/withAsyncValidation';
 export { default as withKiwtFieldArray } from './lib/withKiwtFieldArray';
 
 // Functions, utilities, and misc.
-export { default as customPropertyTypes } from './lib/customPropertyTypes';
 export { default as generateQueryParams } from './lib/generateQueryParams';
 export { default as getSASParams } from './lib/getSASParams';
 export { default as getSiblingIdentifier } from './lib/getSiblingIdentifier';