+++ test/jest/helpers/Harness.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { IntlProvider } from 'react-intl';
-import { QueryClient, QueryClientProvider } from 'react-query'
+import { QueryClient, QueryClientProvider } from 'react-query';
 
 import translations from '../../../translations/stripes-erm-components/en';
 import prefixKeys from './prefixKeys';